LH-3432 : Update import external campaign scores sample file to CSV (#4365)
* Update import external campaign scores sample file to csv

* Handle decimal value in external score import

+++ app/core/admin_jobs/import_external_campaign_scoring.rb
@@ -45,7 +45,7 @@ module AdminJobs
       return unless user
 
       campaign_factor.campaign_factor_values.find_or_initialize_by(user: user, campaign: campaign).tap do |cfv|
-        cfv.value = campaign_factor.numeric_output_type? ? value.to_i : value.to_s
+        cfv.value = campaign_factor.numeric_output_type? ? value.to_d : value.to_s
         cfv.save!
       end
     end
